# optimize wbTransitions