```
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/uniseg
                      │  .old.txt   │              .new.txt               │
                      │   sec/op    │   sec/op     vs base                │
WordFunctionBytes-10    3.038µ ± 1%   2.139µ ± 0%  -29.58% (p=0.000 n=10)
WordFunctionString-10   3.020µ ± 0%   2.113µ ± 0%  -30.05% (p=0.000 n=10)
geomean                 3.029µ        2.126µ       -29.82%

                      │   .old.txt   │              .new.txt               │
                      │     B/op     │    B/op     vs base                 │
WordFunctionBytes-10    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
WordFunctionString-10   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                            ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                      │   .old.txt   │              .new.txt               │
                      │  allocs/op   │ allocs/op   vs base                 │
WordFunctionBytes-10    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
WordFunctionString-10   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                            ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

Before:

![profile001](https://github.com/shogo82148/uniseg/assets/1157344/d1186f8c-09e4-40d1-ace2-b7c3363db331)

After:

![profile002](https://github.com/shogo82148/uniseg/assets/1157344/7f3cb79b-c13b-4b20-a60a-678c3dc3f48c)